Split up actions.js
